Add dovi_parse_rpu_bin_file C function
And refactor C API modules
