+++ src/tests/rpu.rs
@@ -268,9 +268,6 @@ fn profile8_001_end_crc32() -> Result<()> {
     use dolby_vision::rpu::utils::parse_rpu_file;
 
     let rpus = parse_rpu_file(&PathBuf::from("./assets/tests/p8_001_end_crc32.bin"))?;
-    assert!(rpus.is_some());
-
-    let rpus = rpus.unwrap();
     assert_eq!(rpus.len(), 3);
 
     let dovi_rpu = &rpus[0];
