+++ tests/hevc/inject_rpu.rs
@@ -201,7 +201,7 @@ fn duplicated_end() -> Result<()> {
 
     output_rpu.assert(predicate::path::is_file());
 
-    let rpus = dolby_vision::rpu::utils::parse_rpu_file(output_rpu.as_ref())?.unwrap();
+    let rpus = dolby_vision::rpu::utils::parse_rpu_file(output_rpu.as_ref())?;
     assert_eq!(rpus.len(), 259);
 
     Ok(())
